Factor out geo-autocomplete.tsx.
